[FIX] test_website, *: prevent external video call during bg video tour
*: web_editor, website

For some reason, the tour introduced by [1] started failing on the
runbot for its 17.0 version, showcasing turnstile errors. Not sure why
but this tour should never have been making external calls to video URL
anyway and solving that seems to solve the main issue.

Specific turnstile problems will be investigated later.

[1]: https://github.com/odoo/odoo/commit/2789b50bc72cc801fbd5005f74e2cd2655040cbb

runbot-107885

X-original-commit: adbc3aaa552a052705fd3f0afbcfb2595ebb37e7
