+++ addons/test_website/tests/__init__.py
@@ -19,5 +19,6 @@ from . import test_reset_views
 from . import test_restricted_editor
 from . import test_session
 from . import test_settings
+from . import test_snippet_background_video
 from . import test_views_during_module_operation
 from . import test_website_controller_page
