+++ addons/website/tests/test_ui.py
@@ -696,6 +696,3 @@ class TestUi(odoo.tests.HttpCase):
 
     def test_snippet_visibility_option(self):
         self.start_tour("/", "snippet_visibility_option", login="admin")
-
-    def test_snippet_background_video(self):
-        self.start_tour("/", "website_snippet_background_video", login="admin")
